+++ Source/AnimationDescription.swift
@@ -26,10 +26,10 @@ final class AnimationDescription {
     lazy var animation: Animation = {
         Animation(
             keyPath: keyPath,
-            duration: duration,
-            delay: delay,
-            fromValue: fromValue,
-            toValue: toValue,
+            duration: duration.timeValue(),
+            delay: delay.timeValue(),
+            fromValue: fromValue.animationValueFor(animationAttribute: attribute),
+            toValue: toValue.animationValueFor(animationAttribute: attribute),
             easingType: easingType,
             path: path
         )