+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 ## Added
 * Documentation on [Github Pages](lotus-docs.github.io)
 
+## Changed
+* All easings became math functions
+* Custom easing use math function instead of CAMediaTimingFunction
+
 # Changelog for Lotus 0.1.0
 ## Added
 * Tests