+++ CHANGELOG.md
@@ -6,6 +6,9 @@
 * All easings became math functions
 * Custom easing use math function instead of CAMediaTimingFunction
 
+## Fixed
+* [LI-37 Type issues](https://github.com/lotus-ios/lotus/issues/37)
+
 # Changelog for Lotus 0.1.0
 ## Added
 * Tests